New Polymorph and Fox with emote sounds
JuiceThatMakesYouBark needs its own custom polymorph, I was recommended to make it a fox! <3
The code for polymorph.yml is copied from the ArtifactLizard polymorph and edited.

I also added a new entity, MobFoxBarkJuice as a placeholder - though it's probably better if we merge the changes to its parent, MobFox
so that way all foxes can emote properly with the new emoteSounds, UnisexFox